requirements_webapp.txt
streamlit==1.40.2
torch==2.7.1
torchvision==0.22.1
pillow==12.0.0
opencv-python==4.12.0.88
numpy==2.2.6
streamlit-webrtc
av
